requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0

# Data analysis and manipulation
pandas==2.0.3
numpy==1.24.4

# Financial data APIs
yfinance==0.2.18
requests==2.31.0

# Web scraping
beautifulsoup4==4.12.2
feedparser==6.0.10

# Statistical analysis (optional - will fallback if not available)
scikit-learn==1.3.0

# Configuration and environment
python-dotenv==1.0.0

# HTTP client
httpx==0.25.0
pydantic==2.4.2

# Date/time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3